+++ src/pymsx/api/__init__.py
@@ -1,1 +1,0 @@
-"""Module for classes managing api access to remote msx server."""