+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 <a id='changelog-0.0.1'></a>
 # 0.0.1 — 2023-02-20
 
-- Class MsxClient to connect and validate a remote msx server
+- Class MoaiClient to connect and validate a remote moai server
 - Custom exceptions for common errors during authentication and validation
 - Dataclass schemas around authentication and validation
 - Tests for current functionality